refactor: 절기 등록 로직과 스케줄링 로직을 분리
